Add git recommendations and update README.md
The recommendations should help to get started with Git as a version
control system.
It is best practice to set a global ignore file for OS files, which
should not be commited to version control systems. A Git commit
message template provides helping text for writing useful commit text.
In recent Git versions (2.28+) it is possible to set the default
branch (in this case 'main' instead of 'master'). Simple push method
ensures that only the current branch is pushed.